Force UTF-8 encoding.
The HTTPrb library will call #to_json on any object that is getting sent
as JSON. This library uses the JSON library from stdlib which will fail
on encoding errors. Since we want to be resiliant to failures, we encode
into UTF-8 and replace characters if necessary. This isn't strictly
needed for LogDNA, but invalid characters will look bad in LogDNA's UI
regardless.